DenseChangelog: Account for possibility of a folder in the response
Goo.im recently changed their schema to include a nightlies folder.
To account for this change we need a second date check.  If the
first response does not contain "modified" key we assume it is
not a file and check the second object in the JSONArray response.
Should we fail a second time its fatal.

PatchSet 2: Account for missing end date.

Change-Id: I22210d4b17374776658b9a5421d8cfc51b9bc4b4